requirements.txt
alembic==1.5.2
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.0
flask-loguru==1.0.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.3
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
PDFNetPython3==8.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
